RAG_TEST/Tests/LLM_judge/requirements.txt
openai==2.21.0
deepeval==3.8.8
python-dotenv==1.2.2
